fix(streaming): scope replaceable closes to Codex
