+++ tests/integration/streaming/test_resident_drain_manager.py
@@ -10,42 +10,24 @@ from meridian.lib.bootstrap.services import prepare_for_runtime_write
 from meridian.lib.core.context import RuntimeContext
 from meridian.lib.core.types import HarnessId, SpawnId
 from meridian.lib.harness.common import extract_codex_report
-from meridian.lib.harness.semantics import TerminalEventOutcome
 from meridian.lib.ops.spawn.models import SpawnSignalInput
 from meridian.lib.state import spawn_store
 from meridian.lib.state.artifact_store import LocalStore
-from meridian.lib.streaming import descendant_evidence as descendant_evidence_module
 from meridian.lib.streaming.drain_policy import (
-    DrainAction,
     PersistentDrainPolicy,
 )
-from meridian.lib.streaming.resident_drain import ResidentDrainCoordinator
 from tests.support.async_determinism import (
     AsyncDeterminism,
     assert_still_pending,
     wait_until,
 )
-from tests.support.fakes import FakeClock
 from tests.support.pi import start_row
 from tests.support.resident_drain import (
     FakeResidentConnection,
-    awaiting_done_coordinator,
-    coordinator_with_clock,
-    descendant_cancellation_from_roots,
     next_turn_boundary,
     resident_event,
     start_manager,
 )
-
-
-async def _execute_latched_cleanup(
-    coordinator: ResidentDrainCoordinator,
-    outcome: TerminalEventOutcome | None,
-) -> None:
-    exit_decision = await coordinator.handle_stream_exit(outcome)
-    request = exit_decision.post_publication_cleanup
-    assert request is not None
-    await coordinator.execute_post_publication_cleanup(request)
 
 
 @pytest.mark.asyncio
@@ -65,7 +47,6 @@ async def test_codex_terminal_success_without_live_children_finalizes_immediatel
         assert outcome.status == "succeeded"
     finally:
         await manager.stop_spawn(spawn_id)
-
 
 @pytest.mark.asyncio
 async def test_codex_backend_death_with_pending_success_preserves_resident_reason(
@@ -91,6 +72,35 @@ async def test_codex_backend_death_with_pending_success_preserves_resident_reaso
     finally:
         await manager.stop_spawn(spawn_id)
 
+@pytest.mark.asyncio
+async def test_opencode_process_death_with_pending_success_preserves_exit_diagnostics(
+    tmp_path: Path,
+) -> None:
+    spawn_id = SpawnId("p1")
+    start_row(tmp_path, str(spawn_id), HarnessId.OPENCODE, None)
+    start_row(tmp_path, "p2", HarnessId.CODEX, str(spawn_id))
+    connection = FakeResidentConnection(HarnessId.OPENCODE)
+    manager = await start_manager(tmp_path, connection, spawn_id=spawn_id)
+    subscriber = manager.subscribe(spawn_id)
+    assert subscriber is not None
+
+    connection.emit(resident_event(HarnessId.OPENCODE, "session.idle", {}))
+    await next_turn_boundary(subscriber)
+    diagnostic = (
+        "OpenCode subprocess exited with code 17.\n\n"
+        "OpenCode subprocess stderr:\n"
+        "fatal backend detail"
+    )
+    connection.fail_backend(diagnostic)
+
+    try:
+        outcome = await manager.wait_for_completion(spawn_id)
+        assert outcome is not None
+        assert outcome.status == "failed"
+        assert outcome.error == diagnostic
+        assert outcome.error != "backend_dead_while_awaiting_done"
+    finally:
+        await manager.stop_spawn(spawn_id)
 
 @pytest.mark.asyncio
 async def test_opencode_terminal_success_without_live_children_finalizes_immediately(
@@ -110,7 +120,6 @@ async def test_opencode_terminal_success_without_live_children_finalizes_immedia
         assert True not in connection.fake_resident_backend.awaiting_done_values
     finally:
         await manager.stop_spawn(spawn_id)
-
 
 @pytest.mark.asyncio
 async def test_opencode_child_session_idle_does_not_finalize_parent(
@@ -152,7 +161,6 @@ async def test_opencode_child_session_idle_does_not_finalize_parent(
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_opencode_child_session_error_does_not_fail_parent(
     tmp_path: Path,
@@ -189,7 +197,6 @@ async def test_opencode_child_session_error_does_not_fail_parent(
         assert outcome.status == "succeeded"
     finally:
         await manager.stop_spawn(spawn_id)
-
 
 @pytest.mark.parametrize(
     "harness_id,event_type",
@@ -230,7 +237,6 @@ async def test_resident_persistent_policy_emits_boundary_and_stays_alive(
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_opencode_terminal_success_resides_until_child_finishes(
     tmp_path: Path,
@@ -267,7 +273,6 @@ async def test_opencode_terminal_success_resides_until_child_finishes(
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_resident_reconciles_finalizing_child_with_durable_report_as_done(
     tmp_path: Path,
@@ -294,7 +299,6 @@ async def test_resident_reconciles_finalizing_child_with_durable_report_as_done(
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_resident_still_waits_on_genuinely_active_finalizing_child(
     tmp_path: Path,
@@ -320,7 +324,6 @@ async def test_resident_still_waits_on_genuinely_active_finalizing_child(
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_done_signal_at_terminalresident_event_wins_over_outstanding_child(
     tmp_path: Path,
@@ -344,110 +347,6 @@ async def test_done_signal_at_terminalresident_event_wins_over_outstanding_child
         assert True not in connection.fake_resident_backend.awaiting_done_values
     finally:
         await manager.stop_spawn(spawn_id)
-
-
-@pytest.mark.asyncio
-async def test_terminal_done_fails_closed_when_descendant_evidence_unreadable(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from meridian.lib.state.spawn_signals import write_spawn_signal
-    from meridian.lib.streaming import resident_drain as resident_drain_module
-
-    start_row(tmp_path, "p1", HarnessId.CODEX, None)
-    clock = FakeClock(start=100.0)
-    monkeypatch.setattr(resident_drain_module.time, "monotonic", clock.monotonic)
-    connection = FakeResidentConnection(HarnessId.CODEX)
-    coordinator = ResidentDrainCoordinator.for_connection(
-        runtime_root=tmp_path,
-        spawn_id=SpawnId("p1"),
-        receiver=connection,
-        resident_backend=connection.resident_backend,
-        deadline_seconds=30.0,
-        poll_seconds=0.01,
-        rearm_budget=None,
-        cancel_descendants=descendant_cancellation_from_roots(tmp_path, tmp_path),
-    )
-    write_spawn_signal(tmp_path, "p1", "done")
-
-    def _raise_evidence_read_failure(*_args: object) -> None:
-        raise OSError("descendant evidence unavailable")
-
-    monkeypatch.setattr(
-        descendant_evidence_module.spawn_store,
-        "list_spawns",
-        _raise_evidence_read_failure,
-    )
-    terminal = TerminalEventOutcome(status="succeeded", exit_code=0)
-
-    decision = await coordinator.handle_terminal_event(
-        resident_event(HarnessId.CODEX, "turn/completed", {}),
-        terminal,
-        DrainAction(terminate=True, emit_turn_boundary=False),
-    )
-
-    assert decision.recorded_outcome is None
-    assert decision.emit_turn_boundary is True
-    assert connection.fake_resident_backend.awaiting_done_values == [True]
-
-    clock.advance(30.0)
-    expired = await coordinator.handle_timeout()
-
-    assert expired.recorded_outcome is not None
-    assert expired.recorded_outcome.status == "failed"
-    assert expired.recorded_outcome.exit_code == 1
-    assert expired.recorded_outcome.error == "resident_evidence_unreadable"
-
-
-@pytest.mark.asyncio
-async def test_terminal_done_completes_when_descendant_evidence_recovers(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from meridian.lib.state.spawn_signals import write_spawn_signal
-    from meridian.lib.streaming import resident_drain as resident_drain_module
-
-    clock = FakeClock(start=100.0)
-    monkeypatch.setattr(resident_drain_module.time, "monotonic", clock.monotonic)
-    start_row(tmp_path, "p1", HarnessId.CODEX, None)
-    connection = FakeResidentConnection(HarnessId.CODEX)
-    coordinator = ResidentDrainCoordinator.for_connection(
-        runtime_root=tmp_path,
-        spawn_id=SpawnId("p1"),
-        receiver=connection,
-        resident_backend=connection.resident_backend,
-        deadline_seconds=30.0,
-        poll_seconds=0.01,
-        rearm_budget=None,
-        cancel_descendants=descendant_cancellation_from_roots(tmp_path, tmp_path),
-    )
-    evidence_readable = False
-    list_spawns = descendant_evidence_module.spawn_store.list_spawns
-
-    def _read_descendants(runtime_root: Path) -> object:
-        if not evidence_readable:
-            raise OSError("descendant evidence unavailable")
-        return list_spawns(runtime_root)
-
-    monkeypatch.setattr(
-        descendant_evidence_module.spawn_store,
-        "list_spawns",
-        _read_descendants,
-    )
-    write_spawn_signal(tmp_path, "p1", "done")
-    terminal = TerminalEventOutcome(status="succeeded", exit_code=0)
-
-    waiting = await coordinator.handle_terminal_event(
-        resident_event(HarnessId.CODEX, "turn/completed", {}),
-        terminal,
-        DrainAction(terminate=True, emit_turn_boundary=False),
-    )
-    evidence_readable = True
-    recovered = await coordinator.handle_timeout()
-
-    assert waiting.recorded_outcome is None
-    assert recovered.recorded_outcome == terminal
-
 
 @pytest.mark.asyncio
 async def test_spawn_done_op_releases_resident_wait_via_environment_default(
@@ -485,7 +384,6 @@ async def test_spawn_done_op_releases_resident_wait_via_environment_default(
         assert outcome.status == "succeeded"
     finally:
         await manager.stop_spawn(spawn_id)
-
 
 @pytest.mark.asyncio
 async def test_spawn_rearm_op_extends_resident_deadline(
@@ -580,7 +478,6 @@ async def test_spawn_rearm_op_extends_resident_deadline(
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_resident_wait_fans_out_turn_boundary_to_subscriber(tmp_path: Path) -> None:
     spawn_id = SpawnId("p1")
@@ -603,7 +500,6 @@ async def test_resident_wait_fans_out_turn_boundary_to_subscriber(tmp_path: Path
             )
     finally:
         await manager.stop_spawn(spawn_id)
-
 
 @pytest.mark.asyncio
 async def test_child_written_before_terminalresident_event_is_processed_prevents_early_finalize(
@@ -630,7 +526,6 @@ async def test_child_written_before_terminalresident_event_is_processed_prevents
         assert outcome.status == "succeeded"
     finally:
         await manager.stop_spawn(spawn_id)
-
 
 @pytest.mark.asyncio
 async def test_resident_stream_close_with_dead_backend_fails_while_child_running(
@@ -666,7 +561,6 @@ async def test_resident_stream_close_with_dead_backend_fails_while_child_running
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_resident_stream_close_with_stalled_backend_is_not_dead_outcome(
     tmp_path: Path,
@@ -700,7 +594,6 @@ async def test_resident_stream_close_with_stalled_backend_is_not_dead_outcome(
         assert outcome.error == "stream_closed_while_awaiting_done"
     finally:
         await manager.stop_spawn(spawn_id)
-
 
 @pytest.mark.asyncio
 async def test_codex_resident_deadline_waits_then_reaps_live_child(
@@ -773,7 +666,6 @@ async def test_codex_resident_deadline_waits_then_reaps_live_child(
     finally:
         await manager.stop_spawn(spawn_id)
 
-
 @pytest.mark.asyncio
 async def test_codex_resident_finalization_preserves_artifact_report(tmp_path: Path) -> None:
     spawn_id = SpawnId("p1")
@@ -813,202 +705,3 @@ async def test_codex_resident_finalization_preserves_artifact_report(tmp_path: P
         )
     finally:
         await manager.stop_spawn(spawn_id)
-
-
-@pytest.mark.asyncio
-async def test_active_followup_turn_stays_resident_honors_done_and_defers_poll(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from meridian.lib.state.spawn_signals import write_spawn_signal
-    from meridian.lib.streaming import resident_drain as resident_drain_module
-
-    clock = FakeClock(start=100.0)
-    monkeypatch.setattr(resident_drain_module.time, "monotonic", clock.monotonic)
-    start_row(tmp_path, "p1", HarnessId.CODEX, None)
-    start_row(tmp_path, "p2", HarnessId.CODEX, "p1")
-    connection = FakeResidentConnection(HarnessId.CODEX)
-    coordinator = ResidentDrainCoordinator.for_connection(
-        runtime_root=tmp_path,
-        spawn_id=SpawnId("p1"),
-        receiver=connection,
-        resident_backend=connection.resident_backend,
-        deadline_seconds=3300.0,
-        poll_seconds=5.0,
-        rearm_budget=None,
-        cancel_descendants=descendant_cancellation_from_roots(tmp_path, tmp_path),
-    )
-    terminal = TerminalEventOutcome(status="succeeded", exit_code=0)
-    waiting = await coordinator.handle_terminal_event(
-        resident_event(HarnessId.CODEX, "turn/completed", {}),
-        terminal,
-        DrainAction(terminate=True, emit_turn_boundary=False),
-    )
-    assert waiting.emit_turn_boundary is True
-    spawn_store.finalize_spawn(tmp_path, SpawnId("p2"), "succeeded", 0, origin="runner")
-    write_spawn_signal(tmp_path, "p1", "rearm")
-    rearmed = await coordinator.handle_timeout()
-    assert rearmed.recorded_outcome is None
-    assert coordinator.next_timeout() == pytest.approx(5.0)
-    assert connection.fake_resident_backend.injected_messages == []
-
-    clock.advance(270.0)
-    nudged = await coordinator.handle_timeout()
-    assert nudged.recorded_outcome is None
-    assert len(connection.fake_resident_backend.injected_messages) == 1
-    injected = connection.fake_resident_backend.injected_messages[0]
-    assert "meridian spawn done" in injected
-    assert "meridian spawn rearm" in injected
-
-    coordinator.observe_activity_transition("turn_active")
-
-    assert coordinator.next_timeout() == pytest.approx(5.0)
-    assert connection.fake_resident_backend.awaiting_done_values == [True, False]
-
-    write_spawn_signal(tmp_path, "p1", "done")
-    decision = await coordinator.handle_timeout()
-
-    assert decision.recorded_outcome == terminal
-    assert coordinator.next_timeout() is None
-    assert connection.fake_resident_backend.awaiting_done_values == [True, False, False]
-
-
-@pytest.mark.asyncio
-async def test_active_followup_turn_still_enforces_deadline(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from meridian.lib.streaming import resident_drain as resident_drain_module
-
-    clock = FakeClock(start=0.0)
-    monkeypatch.setattr(resident_drain_module.time, "monotonic", clock.monotonic)
-    start_row(tmp_path, "p1", HarnessId.CODEX, None)
-    connection = FakeResidentConnection(HarnessId.CODEX)
-    coordinator = await coordinator_with_clock(tmp_path, connection, clock, deadline_seconds=10.0)
-    coordinator.observe_activity_transition("turn_active")
-
-    clock.advance(10.0)
-    decision = await coordinator.handle_timeout()
-
-    assert decision.recorded_outcome is not None
-    assert decision.recorded_outcome.status == "timed_out"
-    assert decision.recorded_outcome.error == "resident_deadline_expired"
-
-
-@pytest.mark.asyncio
-async def test_deadline_returns_timed_out_when_descendant_reap_fails(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from meridian.lib.bootstrap import services as bootstrap_services
-    from meridian.lib.streaming import resident_drain as resident_drain_module
-
-    clock = FakeClock(start=0.0)
-    monkeypatch.setattr(resident_drain_module.time, "monotonic", clock.monotonic)
-    start_row(tmp_path, "p1", HarnessId.CODEX, None)
-    connection = FakeResidentConnection(HarnessId.CODEX)
-    coordinator = await coordinator_with_clock(tmp_path, connection, clock, deadline_seconds=10.0)
-    cleanup_calls = 0
-
-    class _FailingService:
-        async def cancel_descendants(self, root_id: SpawnId) -> set[str]:
-            nonlocal cleanup_calls
-            _ = root_id
-            cleanup_calls += 1
-            raise RuntimeError("teardown failed")
-
-    monkeypatch.setattr(
-        bootstrap_services,
-        "build_spawn_application_service_from_roots",
-        lambda _project_root, _runtime_root: _FailingService(),
-    )
-    clock.advance(10.0)
-
-    decision = await coordinator.handle_timeout()
-
-    assert decision.recorded_outcome is not None
-    assert decision.recorded_outcome.status == "timed_out"
-    assert decision.recorded_outcome.error == "resident_deadline_expired"
-    assert cleanup_calls == 0
-    await _execute_latched_cleanup(coordinator, decision.recorded_outcome)
-    assert cleanup_calls == 1
-
-    later_decision = await coordinator.handle_timeout()
-    assert later_decision.recorded_outcome is None
-    assert cleanup_calls == 1
-
-
-@pytest.mark.asyncio
-async def test_deadline_reap_cancels_active_descendant_cli_spawns(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from meridian.lib.bootstrap import services as bootstrap_services
-    from meridian.lib.state.spawn_tree import active_descendants
-    from meridian.lib.streaming import resident_drain as resident_drain_module
-
-    clock = FakeClock(start=0.0)
-    monkeypatch.setattr(resident_drain_module.time, "monotonic", clock.monotonic)
-    start_row(tmp_path, "p1", HarnessId.CODEX, None)
-    start_row(tmp_path, "p2", HarnessId.OPENCODE, "p1")
-    start_row(tmp_path, "p3", HarnessId.OPENCODE, "p1")
-    spawn_store.finalize_spawn(tmp_path, SpawnId("p3"), "succeeded", 0, origin="runner")
-    connection = FakeResidentConnection(HarnessId.CODEX)
-    coordinator = await coordinator_with_clock(tmp_path, connection, clock, deadline_seconds=10.0)
-    cancelled: list[str] = []
-
-    class _FakeService:
-        async def cancel_descendants(self, root_id: SpawnId) -> set[str]:
-            reaped_ids: set[str] = set()
-            for descendant in active_descendants(tmp_path, root_id):
-                reaped_ids.add(descendant.id)
-                cancelled.append(descendant.id)
-                spawn_store.finalize_spawn(
-                    tmp_path,
-                    descendant.id,
-                    "cancelled",
-                    130,
-                    origin="cancel",
-                    error="cancelled",
-                )
-            return reaped_ids
-
-    monkeypatch.setattr(
-        bootstrap_services,
-        "build_spawn_application_service_from_roots",
-        lambda _project_root, _runtime_root: _FakeService(),
-    )
-    clock.advance(10.0)
-
-    decision = await coordinator.handle_timeout()
-
-    assert decision.recorded_outcome is not None
-    assert decision.recorded_outcome.status == "timed_out"
-    assert cancelled == []
-    await _execute_latched_cleanup(coordinator, decision.recorded_outcome)
-    assert cancelled == ["p2"]
-    child = spawn_store.get_spawn(tmp_path, SpawnId("p2"))
-    assert child is not None
-    assert child.status == "cancelled"
-    already_terminal = spawn_store.get_spawn(tmp_path, SpawnId("p3"))
-    assert already_terminal is not None
-    assert already_terminal.status == "succeeded"
-
-
-@pytest.mark.asyncio
-async def test_done_signal_is_honored_with_tracked_child_outstanding(
-    tmp_path: Path,
-) -> None:
-    from meridian.lib.state.spawn_signals import write_spawn_signal
-
-    start_row(tmp_path, "p1", HarnessId.OPENCODE, None)
-    start_row(tmp_path, "p2", HarnessId.CODEX, "p1")
-    connection = FakeResidentConnection(HarnessId.OPENCODE)
-    coordinator = await awaiting_done_coordinator(tmp_path, connection)
-
-    write_spawn_signal(tmp_path, "p1", "done")
-    decision = await coordinator.handle_timeout()
-
-    assert decision.recorded_outcome is not None
-    assert decision.recorded_outcome.status == "succeeded"
-    assert connection.fake_resident_backend.injected_messages == []
